Keep paired pane cue coverage in existing UI spec
